Update MSODTransferSpec.md
Data message and Sys message had common fields but stored in different places.  When thinking of the JSON objects with inheritance, we should lump all the same fields in the same place, so we can make a common Message class that Sys Message and Data Message inherent from.   This make serializing and deserializing and configuration easier when we have C# classes that reflect the Messages.